% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{cv_seq}
\alias{cv_seq}
\title{Generate a sequence based on coefficient of variation}
\usage{
cv_seq(point, cv = 30, n = 5, nsd = 2)
}
\arguments{
\item{point}{reference parameter value}

\item{cv}{coefficient of variation}

\item{n}{number of values to simulate in the sequence}

\item{nsd}{number of standard deviations defining the
range of simulated parameter values}
}
\description{
Generate a sequence based on coefficient of variation
}
