% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{even_seq}
\alias{even_seq}
\title{Generate evenly spaced sequence}
\usage{
even_seq(from, to, .n = 5)
}
\arguments{
\item{from}{passed to \code{\link{seq}}}

\item{to}{passed to \code{\link{seq}}}

\item{.n}{passed to \code{\link{seq}} as \code{length.out}}
}
\description{
Generate evenly spaced sequence
}
\examples{
even_seq(1, 10, 4)

}
