% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{fct_seq}
\alias{fct_seq}
\title{Generate a sequence by fold increase and decrease from a point}
\usage{
fct_seq(point, .n = 5, .factor = c(3, 3), .geo = TRUE)
}
\arguments{
\item{point}{a numeric vector of length 1}

\item{.n}{number of elements in the sequence}

\item{.factor}{an integer vector of length 1 or 2; if length 1, 
values will be recycled to length 2; the first number used to divide
`point` to generate the minimum value in the sequence; the second 
number is used to multiply `point` to generate the 
maximum value in the sequence}

\item{.geo}{if `TRUE`, [geo_seq()] is used to generate
the sequence; otherwise, [even_seq()] is used to generate 
the sequence}
}
\description{
Generate a sequence by fold increase and decrease from a point
}
\examples{

fct_seq(10)

}
