% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseq.R
\name{geo_seq}
\alias{geo_seq}
\title{Generate a geometric sequence of parameter values}
\usage{
geo_seq(from, to, .n = 5)
}
\arguments{
\item{from}{passed to \code{\link{seq}}}

\item{to}{passed to \code{\link{seq}}}

\item{.n}{passed to \code{\link{seq}} as \code{length.out}}
}
\description{
Generate a geometric sequence of parameter values
}
\examples{
geo_seq(1,10,10)

}
