% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsens.R
\name{mrgsens}
\alias{mrgsens}
\title{Run ad-hoc parameter sensitivity analyses with mrgsolve}
\usage{
mrgsens(
  mod,
  par,
  method = c("factor", "cv", "range", "manual"),
  type = "each",
  ...
)
}
\arguments{
\item{mod}{a mrgsolve model object}

\item{par}{parameter names for sensitivity analysis}

\item{method}{parameter sequence generation method}

\item{type}{use `each` to vary one parameter at a time}

\item{...}{passed to [sens_each]}
}
\description{
Run ad-hoc parameter sensitivity analyses with mrgsolve
}
