% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\name{mrgsim.sa}
\alias{mrgsim.sa}
\title{Sensitivity Analysis with 'mrgsolve'}
\description{
Perform local sensitivity analysis on ordinary differential
equation based models, including ad-hoc graphical analyses based on
sequences of parameters as well as local sensitivity analysis. Functions
are provided for creating inputs, simulating scenarios and plotting outputs.
}
\details{
\itemize{
\item Local sensitivity analysis: \code{\link[=lsa]{lsa()}}, \code{\link[=lsa_plot]{lsa_plot()}}
\item Run ad-hoc sensitivity analyses: \code{\link[=sens_each]{sens_each()}}, \code{\link[=sens_grid]{sens_grid()}}, \code{\link[=sens_run]{sens_run()}}
\itemize{
\item Use \code{\link[=sens_each_data]{sens_each_data()}} and \code{\link[=sens_grid_data]{sens_grid_data()}} to pass in data sets
}
\item Parameter sequence generation:
\itemize{
\item In a pipeline: \code{\link[=parseq_cv]{parseq_cv()}}, \code{\link[=parseq_fct]{parseq_fct()}}, \code{\link[=parseq_range]{parseq_range()}}, \code{\link[=parseq_manual]{parseq_manual()}}
\item Stand alone: \code{\link[=seq_cv]{seq_cv()}}, \code{\link[=seq_fct]{seq_fct()}}, \code{\link[=seq_geo]{seq_geo()}}, \code{\link[=seq_even]{seq_even()}}
}
\item Plot ad-hoc sensitivity analysis results
\itemize{
\item Use \code{\link[=sens_plot]{sens_plot()}}
}
\item Select parameters or results
\itemize{
\item Use \code{\link[=select_par]{select_par()}} to choose which parameters to vary
\item Use \code{\link[=select_sens]{select_sens()}} to subset sensitivity analysis results
}
}
}
\seealso{
\code{vignette("mrgsim.sa", package = "mrgsim.sa")} for a complete tutorial.
}
