% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseq.R
\name{parseq_factor}
\alias{parseq_factor}
\title{Generate a sequence of parameters}
\usage{
parseq_factor(mod, ..., .n = 5, .factor = 2, .geo = TRUE)
}
\arguments{
\item{mod}{a model object}

\item{...}{unquoted parameter names}

\item{.n}{number of parameters to simulate between the minimum
and maximum parameter values}

\item{.factor}{a numeric vector used to divide and multiply the 
parameter value thus generating the minimum and maximum parameter values,
respectively, for the sequence; if `.factor` is length 1 it will be 
recycled to length 2; the first value is used to divide the nominal 
value generating the minimum value; the second value is used to multiply
the nominal value generating the maximum value}

\item{.geo}{if `TRUE` a geometric sequence is generated (evenly
spaced from min to max on log scale); otherwise, the sequence is 
evenly spaced on cartesian scale}
}
\description{
Generate a sequence of parameters
}
