% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseq.R
\name{parseq_reference}
\alias{parseq_reference}
\title{Set reference values for each parameter}
\usage{
parseq_reference(mod, ..., .ref = list(), .auto = TRUE)
}
\description{
Set reference values for each parameter
}
