% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sens.R
\name{plot_sens}
\alias{plot_sens}
\title{Plot the results of a sensitivity analysis}
\usage{
plot_sens(data, y, x = "time", col = split, log = FALSE, split = FALSE,
  lwd = 0.75)
}
\arguments{
\item{data}{simulated data}

\item{y}{the column to use for the y-axis}

\item{x}{the column to use for the x-axis}

\item{col}{logical; if \code{TRUE} the plot will be colored by 
the parameter values}

\item{log}{if \code{TRUE}, the y-axis will be presented on log scale}

\item{split}{if \code{TRUE}, a list of plots will be returned, one 
for each parameter in the sensitivity analysis}

\item{lwd}{passed to \code{\link[ggplot2]{geom_line}}}
}
\description{
Plot the results of a sensitivity analysis
}
