% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_par}
\alias{select_par}
\title{Identify parameters in a model for sensitivity analysis}
\usage{
select_par(mod, ...)
}
\arguments{
\item{mod}{an mrgsolve model object.}

\item{...}{unquoted parameter names.}
}
\value{
The model object with the selected parameters stored for use by
\code{\link[=parseq_fct]{parseq_fct()}}, \code{\link[=parseq_cv]{parseq_cv()}}, \code{\link[=parseq_range]{parseq_range()}}, or \code{\link[=parseq_manual]{parseq_manual()}}.
}
\description{
Identify parameters in a model for sensitivity analysis
}
\examples{
mod <- mrgsolve::house()
select_par(mod, CL, VC)

}
