% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_each.R
\name{select_sens}
\alias{select_sens}
\title{Select sensitivity runs from a sens_each object}
\usage{
select_sens(x, dv_name = NULL, p_name = NULL)
}
\arguments{
\item{x}{a sens_each object}

\item{dv_name}{character names of dependent variables to select}

\item{p_name}{character names of parameters to select}
}
\description{
Select sensitivity runs from a sens_each object
}
