% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_each.R
\name{as.data.frame.sens_each}
\alias{as.data.frame.sens_each}
\title{Coerce sens_each output to data frame}
\usage{
\method{as.data.frame}{sens_each}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{.case}{integer scenario number}
}
\description{
Coerce sens_each output to data frame
}
