% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_each.R
\name{denest}
\alias{denest}
\title{Unnest a sens_each object}
\usage{
denest(x, keep_id = FALSE)
}
\arguments{
\item{x}{a sens_each object.}

\item{keep_id}{if \code{FALSE} then the \code{ID} column is removed .}
}
\value{
\code{x} is returned after unnesting and possible modification.
}
\description{
Unnest a sens_each object
}
\keyword{internal}
