% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{geo_seq}
\alias{geo_seq}
\title{Generate a geometric sequence of parameter values}
\usage{
geo_seq(from, to, n = 5)
}
\arguments{
\item{from}{passed to [base::seq()]}

\item{to}{passed to [base::seq()]}

\item{n}{passed to [base::seq()] as `length.out`}
}
\description{
Generate a geometric sequence of parameter values
}
\examples{
geo_seq(1,10,10)

}
