% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsens.R
\name{mrgsens}
\alias{mrgsens}
\title{Run ad-hoc parameter sensitivity analyses with mrgsolve}
\usage{
mrgsens(
  mod,
  par,
  method = c("factor", "cv", "range", "manual"),
  type = "each",
  ...
)
}
\description{
Run ad-hoc parameter sensitivity analyses with mrgsolve
}
