% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseq.R
\name{parseq_cv}
\alias{parseq_cv}
\title{Generate a sequence of parameters based on CV}
\usage{
parseq_cv(mod, ..., .cv = 30, .n = 5, .nsd = 2)
}
\arguments{
\item{mod}{a model object}

\item{...}{model parameter names}

\item{.cv}{a coefficient of variation used to determine
range of test parameters}

\item{.n}{number of parameters to simulate in the sequence}

\item{.nsd}{number of standard deviations used to determine the range}
}
\description{
Generate a sequence of parameters based on CV
}
\details{
\itemize{
\item \code{.cv}  is passed to \code{\link[=cv_seq]{cv_seq()}} as \code{cv}
\item \code{.n}   is passed to \code{\link[=cv_seq]{cv_seq()}} as \code{n}
\item \code{.nsd} is passed to \code{\link[=cv_seq]{cv_seq()}} as \code{nsd}
}
}
\seealso{
\code{\link[=parseq_factor]{parseq_factor()}}, \code{\link[=parseq_range]{parseq_range()}}, \code{\link[=parseq_manual]{parseq_manual()}}
}
