% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseq.R
\name{parseq_range}
\alias{parseq_range}
\title{Simulation helper to generate a sequence of parameters from a range}
\usage{
parseq_range(mod, ..., .n = 5, .geo = TRUE)
}
\arguments{
\item{mod}{mrgsolve model object}

\item{...}{unquoted parameter names,}

\item{.n}{number of values to simulate for each parameter sequence}

\item{.geo}{if \code{TRUE} generate a geometric sequence; otherwise,
generate a sequence evenly spaced on Cartesian scale; see \code{\link[=geo_seq]{geo_seq()}}}
}
\description{
Simulation helper to generate a sequence of parameters from a range
}
\details{
\itemize{
\item \code{.n}  is passed to \code{\link[=geo_seq]{geo_seq()}} as \code{n}
}
}
\seealso{
\code{\link[=parseq_cv]{parseq_cv()}}, \code{\link[=parseq_factor]{parseq_factor()}}, \code{\link[=parseq_manual]{parseq_manual()}}
}
