% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseq.R
\name{parseq_reference}
\alias{parseq_reference}
\title{Set reference values for each parameter}
\usage{
parseq_reference(mod, auto = TRUE)
}
\arguments{
\item{mod}{a model object}

\item{auto}{if \code{TRUE} then the model parameter list is used}
}
\description{
Set reference values for each parameter
}
