% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsa.R
\name{plot.lsa}
\alias{plot.lsa}
\title{Plot a lsa object}
\usage{
\method{plot}{lsa}(x, y = NULL, pal = NULL, ...)
}
\arguments{
\item{x}{output from \code{\link[=lsa]{lsa()}}.}

\item{y}{not used.}

\item{pal}{a color palette passed to \code{\link[ggplot2:scale_brewer]{ggplot2::scale_color_brewer()}}; use
\code{NULL} to use default ggplot color scale.}

\item{...}{not used.}
}
\value{
A ggplot.
}
\description{
Plot a lsa object
}
\keyword{internal}
