% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vera.R
\name{plot.lsa}
\alias{plot.lsa}
\title{Plot a lsa object}
\usage{
\method{plot}{lsa}(x, y = NULL, ...)
}
\arguments{
\item{x}{output from \link{lsa}}

\item{y}{not used}

\item{...}{not used}
}
\description{
Plot a lsa object
}
