% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_par}
\alias{select_par}
\title{Identify parameters in a model for sensitivity analysis}
\usage{
select_par(mod, ...)
}
\arguments{
\item{mod}{an mrgsolve model object.}

\item{...}{unquoted parameter names.}
}
\description{
Identify parameters in a model for sensitivity analysis
}
\examples{
mod <- mrgsolve::house()
select_par(mod, CL, VC)

}
