% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens.R
\name{sens_each}
\alias{sens_each}
\alias{sens_grid}
\title{Run a sensitivity analysis}
\usage{
sens_each(mod, idata = NULL, ...)

sens_grid(mod, idata = NULL, ...)
}
\arguments{
\item{mod}{a model object}

\item{idata}{included only to prevent users from passing through; the 
function will create an idata set if appropriate}

\item{...}{passed to \code{\link[mrgsolve]{mrgsim_df}}}
}
\description{
Use \code{sens_each} to examine sequences of parameters one 
at a time.  Use \code{sens_grid} to examine all combinations of 
sequences of parameters.
}
