% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens.R, R/sens_each.R, R/sens_grid.R
\name{sens_fun}
\alias{sens_fun}
\alias{sens_each}
\alias{sens_each_data}
\alias{sens_grid}
\alias{sens_grid_data}
\title{Run a sensitivity analysis}
\usage{
sens_each(mod, idata = NULL, ...)

sens_each_data(mod, data, idata = NULL, ...)

sens_grid(mod, idata = NULL, ...)

sens_grid_data(mod, data, idata = NULL, ...)
}
\arguments{
\item{mod}{a model object}

\item{idata}{included only to prevent users from passing through; the 
function will create an idata set if appropriate}

\item{...}{passed to \code{\link[mrgsolve]{mrgsim_df}}}

\item{data}{a simulation input data set (see \code{\link[mrgsolve]{data_set}})}
}
\description{
Use \code{sens_each} to examine sequences of parameters one 
at a time.  Use \code{sens_grid} to examine all combinations of 
sequences of parameters.
}
