% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseq.R
\name{sens_fun}
\alias{sens_fun}
\alias{sens_each}
\alias{sens_fun}
\alias{sens_grid}
\title{Run a sensitivity analysis}
\usage{
sens_each(mod, ...)

sens_grid(mod, ...)
}
\arguments{
\item{mod}{a model object}

\item{...}{passed to \code{\link[mrgsolve]{mrgsim_df}}}
}
\description{
Use \code{sens_each} to examine sequences of parameters one 
at a time.  Use \code{sens_grid} to examine all combinations of 
sequences of parameters.
}
