% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sens.R
\name{sens_plot}
\alias{sens_plot}
\alias{sens_plot.sens_each}
\alias{sens_plot.sens_grid}
\title{Plot sensitivity analysis results}
\usage{
sens_plot(data, ...)

\method{sens_plot}{sens_each}(data, col, log = FALSE, ncol = NULL, ...)

\method{sens_plot}{sens_grid}(data, col, digits = 2, ncol = NULL, ...)
}
\arguments{
\item{data}{output from \code{\link{sens_each}} or 
\code{\link{sens_grid}}}

\item{...}{arguments passed on to methods}

\item{col}{output column name to plot}

\item{log}{if \code{TRUE}, y-axis is transformed to log scale}

\item{ncol}{passed to \code{\link[ggplot2]{facet_wrap}}}

\item{digits}{used to format numbers on the strips}
}
\description{
Plot sensitivity analysis results
}
