% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sens.R
\name{sens_plot}
\alias{sens_plot}
\alias{sens_plot.sens_each}
\alias{sens_plot.sens_grid}
\title{Plot sensitivity analysis results}
\usage{
sens_plot(data, ...)

\method{sens_plot}{sens_each}(
  data,
  dv_name,
  logy = FALSE,
  ncol = NULL,
  bw = FALSE,
  digits = 3,
  plot_ref = TRUE,
  xlab = "time",
  ylab = dv_name[1],
  cowplot = TRUE,
  ...
)

\method{sens_plot}{sens_grid}(data, dv_name, digits = 2, ncol = NULL, logy = FALSE, ...)
}
\arguments{
\item{data}{output from [sens_each()] or 
[sens_grid()]}

\item{...}{arguments passed on to methods}

\item{dv_name}{output column name to plot}

\item{logy}{if `TRUE`, y-axis is transformed to log scale}

\item{ncol}{passed to [ggplot2::facet_wrap()]}

\item{bw}{if `TRUE` a simple black and white plot will be generated
when the [sens_each()] method is used}

\item{digits}{used to format numbers on the strips}

\item{plot_ref}{if `TRUE`, then the reference case will be plotted in a black
dashed line}

\item{xlab}{x-axis title}

\item{ylab}{y-axis title}

\item{cowplot}{if `TRUE`, plots from the `sens_each` method
will be passed through [cowplot::plot_grid()]}
}
\description{
Plot sensitivity analysis results
}
