% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sens.R
\name{sens_plot}
\alias{sens_plot}
\alias{sens_plot.sens_each}
\alias{sens_plot.sens_grid}
\title{Plot sensitivity analysis results}
\usage{
sens_plot(data, ...)

\method{sens_plot}{sens_each}(
  data,
  dv_name,
  logy = FALSE,
  ncol = NULL,
  bw = FALSE,
  digits = 3,
  plot_ref = TRUE,
  xlab = "time",
  ylab = dv_name[1],
  cowplot = TRUE,
  ...
)

\method{sens_plot}{sens_grid}(data, col, digits = 2, ncol = NULL, ...)
}
\arguments{
\item{data}{output from \code{\link{sens_each}} or 
\code{\link{sens_grid}}}

\item{...}{arguments passed on to methods}

\item{dv_name}{output column name to plot}

\item{logy}{if \code{TRUE}, y-axis is transformed to log scale}

\item{ncol}{passed to \code{\link[ggplot2]{facet_wrap}}}

\item{bw}{if \code{TRUE} a simple black and white plot will be generated
when the \code{sens_each} method is used}

\item{digits}{used to format numbers on the strips}

\item{plot_ref}{if `TRUE`, then the reference case will be plotted in a black
dashed line}

\item{xlab}{x-axis title}

\item{ylab}{y-axis title}

\item{cowplot}{if \code{TRUE}, plots from the \code{sens_each} method
will be passed through \code{cowplot::plot_grid()}}

\item{col}{outputs to plot}
}
\description{
Plot sensitivity analysis results
}
