% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{seq_cv}
\alias{seq_cv}
\title{Generate a sequence based on coefficient of variation}
\usage{
seq_cv(point, cv = 30, n = 5, nsd = 2, digits = NULL)
}
\arguments{
\item{point}{reference parameter value.}

\item{cv}{coefficient of variation.}

\item{n}{number of values to simulate in the sequence.}

\item{nsd}{number of standard deviations defining the range of simulated \
parameter values.}

\item{digits}{number of significant digits in the answer; if \code{NULL} (the
default) all digits are retained.}
}
\description{
Generate a sequence based on coefficient of variation
}
\examples{
seq_cv(10)

seq_cv(5, n = 10)

}
