% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{seq_even}
\alias{seq_even}
\title{Generate evenly spaced sequence}
\usage{
seq_even(from, to, n = 5, digits = NULL)
}
\arguments{
\item{from}{passed to \code{\link[base:seq]{base::seq()}}.}

\item{to}{passed to \code{\link[base:seq]{base::seq()}}.}

\item{n}{passed to \code{\link[base:seq]{base::seq()}} as \code{length.out}.}

\item{digits}{number of significant digits in the answer; if \code{NULL} (the
default) all digits are retained.}
}
\description{
Generate evenly spaced sequence
}
\examples{
seq_even(1, 10, 4)

}
