% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{seq_geo}
\alias{seq_geo}
\title{Generate a geometric sequence of parameter values}
\usage{
seq_geo(from, to, n = 5, digits = NULL)
}
\arguments{
\item{from}{passed to \code{\link[base:seq]{base::seq()}}}

\item{to}{passed to \code{\link[base:seq]{base::seq()}}}

\item{n}{passed to \code{\link[base:seq]{base::seq()}} as \code{length.out}}

\item{digits}{number of significant digits in the answer; if \code{NULL} (the
default) all digits are retained}
}
\description{
Generate a geometric sequence of parameter values
}
\examples{
seq_geo(1,10,10)

}
